pyproject.toml
[project]
name = "elt-pipeline-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Rebecca Raezada",email = "rebecca.raezada22@gmail.com"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "requests (>=2.33.0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "anthropic (>=0.86.0,<0.87.0)",
    "dbt-postgres (>=1.10.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
